infer underlays from acceptors and ejectors
