+++ src/js/game/map_chunk_view.js
@@ -50,7 +50,8 @@ export class MapChunkView extends MapChunk {
             systems.mapResources.drawChunk(parameters, this);
         }
 
-        systems.beltUnderlays.drawChunk(parameters, this);
+        systems.acceptorBelt.drawChunk(parameters, this);
+        systems.ejectorBelt.drawChunk(parameters, this);
         systems.belt.drawChunk(parameters, this);
     }
